requirements.txt
spacy
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl
transformers
sentence-transformers
pytesseract
pdfplumber
pillow
faiss-cpu
streamlit
accelerate
torch
numpy
pandas
requests
